Fix /api/authors/* for http response codes

+++ api/authors/delete.php
@@ -19,14 +19,21 @@ $author = new Author($db);
 // Get raw json data from request
 $data = json_decode(file_get_contents('php://input'));
 
-// Set the object property - MUST contain the id
-$author->set_id($data->id);
+// Make sure raw json data includes 'id'
+if(isset($data->id)) {
+    // Set the object property - MUST contain the id
+    $author->set_id($data->id);
 
-// Delete Author
-if($author->delete()) {
-    echo json_encode(array('message' => 'Author Deleted'));
-    http_response_code(200);        // 200 OK
+    // Delete Author
+    if($author->delete()) {
+        echo json_encode(array('message' => 'Author Deleted'));
+        http_response_code(200);        // 200 OK
+    } else {
+        echo json_encode(array('message' => 'Author Not Deleted'));
+        http_response_code(404);        // 404 Not Found
+    }
+
 } else {
-    echo json_encode(array('message' => 'Author Not Deleted'));
-    http_response_code(404);        // 404 Not Found
+    echo json_encode(array('message' => 'Author Not Deleted. MUST contain \'id\'.'));
+    http_response_code(400);        // 400 Bad Request
 }
